Update mixins.less
comments in the varibles used in this file. corrects a bug in margin parametric mixin. refactors button mixin variables  to be more intuitive for button actions. -- we want to minimaize the colors used to maximize the impact and emotional connection purity for the user. 
a color for the containing div's background. an opposing color for the button background so the use sees it right. 
from here the buttons fontcolor can be the same as the container div's background color --giving the button a lookthrough feel on the letters. 
when user clicks the button the font and backgoruns color of button flip while shadows happend and boom next screen.